mental-health-chatbot/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.1
langchain>=0.0.267
openai>=1.0.0
pinecone-client>=2.2.1
pydantic>=1.9.0
python-multipart>=0.0.5
tiktoken>=0.4.0
sentence-transformers>=2.2.2
pypdf>=3.11.0
faiss-cpu>=1.7.4
